+++ sdk/schemaregistry/schema-registry/src/index.ts
@@ -3,13 +3,20 @@
 
 export { SchemaRegistryClient } from "./schemaRegistryClient.js";
 export {
+  type GetAllSchemasOptions,
   type GetSchemaOptions,
   type GetSchemaPropertiesOptions,
   KnownSchemaFormats,
+  type ListSchemasOptions,
   type RegisterSchemaOptions,
   type Schema,
   type SchemaDescription,
   type SchemaRegistry,
   type SchemaRegistryClientOptions,
   type SchemaProperties,
+  type SchemaSummary,
+  SchemaSortOrder,
 } from "./models.js";
+
+/** @internal */
+export { prepareSchemaContent } from "./operations.js";